fix(service): 일관적으로 입,출력에서 Long 사용하도록 변경

+++ src/main/kotlin/kr/hhplus/be/server/service/order/service/OrderService.kt
@@ -46,7 +46,7 @@ class OrderService(
     ) {
         data class ProductWithQuantity(
             val productId: Long,
-            val quantity: Int
+            val quantity: Long
         )
     }
 
@@ -118,7 +118,7 @@ class OrderService(
             val product: Product = sale.product
             val productId: Long = product.requiresId()
 
-            val quantity: Int = sale.soldCount
+            val quantity: Long = sale.soldCount
             val now: ZonedDateTime = context.now
